feat(framework): extend the test package rule to integration tests via path mirroring
