--- tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/CoversPackageMatchRule/Covered/UnpackagedService.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\CoversPackageMatchRule\Covered;
-
-class UnpackagedService
-{
-}
