+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/Covered/CheckoutHelperTrait.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\CoversPackageMatchRule\Covered;
+namespace Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\TestPackageMatchRule\Covered;
 
 use Shopware\Core\Framework\Log\Package;
