+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/Tests/CommercialSuiteFixture.php
@@ -1,11 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Commercial\Tests\Unit\DevOps\StaticAnalyse\PHPStan\Rules\data\CoversPackageMatchRule;
+namespace Shopware\Commercial\Tests\Unit\DevOps\StaticAnalyse\PHPStan\Rules\data\TestPackageMatchRule;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\CoversPackageMatchRule\Covered\CheckoutService;
+use Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\TestPackageMatchRule\Covered\CheckoutService;
 
 #[Package('framework')]
 #[CoversClass(CheckoutService::class)]
