+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/Tests/IntegrationFundamentalsFixture.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Core\System\Locale;
+
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Log\Package;
+
+#[Package('discovery')]
+class IntegrationFundamentalsFixture extends TestCase
+{
+}
