+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/Tests/IntegrationMatchingFixture.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Core\Checkout\Cart;
+
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Log\Package;
+
+#[Package('checkout')]
+class IntegrationMatchingFixture extends TestCase
+{
+}
