+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/Tests/IntegrationUnmirroredFixture.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Recovery\Install;
+
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Log\Package;
+
+#[Package('framework')]
+class IntegrationUnmirroredFixture extends TestCase
+{
+}
