+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/project/src/Core/Checkout/Cart/CartService.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\TestPackageMatchRule\project\src\Core\Checkout\Cart;
+
+use Shopware\Core\Framework\Log\Package;
+
+#[Package('checkout')]
+class CartService
+{
+}
