+++ tests/devops/Core/DevOps/StaticAnalyse/PHPStan/Rules/data/TestPackageMatchRule/project/src/Core/System/Locale/LocaleFormatter.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\DevOps\Core\DevOps\StaticAnalyse\PHPStan\Rules\data\TestPackageMatchRule\project\src\Core\System\Locale;
+
+use Shopware\Core\Framework\Log\Package;
+
+#[Package('fundamentals@discovery')]
+class LocaleFormatter
+{
+}
